requirements.txt
numpy==1.21.5
pandas==1.4.2
python_highcharts==0.4.2
streamlit==1.12.2
